feat: Move DownloadAndInstaller to kmpuiviews and Update Dependencies
This commit moves the `DownloadAndInstaller` to the `kmpuiviews` module for multiplatform use, cleans up related dependencies, and prepares for further multiplatform migration.

-   **DownloadAndInstaller Migration:**
    - Moved `DownloadAndInstaller` to `kmpuiviews`.
    - Removed the old `DownloadAndInstaller` from `UIViews`.
    - Updated `ExtensionListViewModel` to use the new `DownloadAndInstaller` from `kmpuiviews`.
    - Updated `DownloadAndInstallWorker` to use the new `DownloadAndInstaller` from `kmpuiviews`.
    - Updated `PrereleaseViewModel` to use `DownloadAndInstallStatus` from `kmpuiviews`.
    - Updated `PrereleaseScreen` to use the `DownloadAndInstallStatus` from `kmpuiviews`.
    - Updated `MoreInfoViewModel` to use the new `DownloadAndInstaller` from `kmpuiviews`.
    - Updated `DownloadStateRepository` to no longer use `DownloadAndInstaller`.
    - Updated `DownloadStateScreen` to use the `DownloadAndInstallStatus` from `kmpuiviews`.
- **Other**
   - Created a new file to handle `DownloadAndInstaller` on android.
   - Created a new koin module for `DownloadAndInstaller` on android.
   - Removed `DownloadAndInstaller` from the `UIViews` di module
   - Updated `OnboardingContent` to mention that some components can be moved to `kmpuiviews`.
   - Fixed some typos in the comments in `TvUtils`.
   - Removed the update check from the tv side for now.
   - General cleanup.
