+++ UIViews/src/main/java/com/programmersbox/uiviews/presentation/settings/downloadstate/DownloadStateRepository.kt
@@ -3,6 +3,7 @@ package com.programmersbox.uiviews.presentation.settings.downloadstate
 import android.content.Context
 import androidx.core.net.toUri
 import androidx.work.WorkManager
+import com.programmersbox.kmpuiviews.utils.DownloadAndInstaller
 import com.programmersbox.uiviews.checkers.DownloadAndInstallWorker
 import com.programmersbox.uiviews.checkers.DownloadWorker
 import java.io.File
