+++ UIViews/src/main/java/com/programmersbox/uiviews/presentation/settings/downloadstate/DownloadStateScreen.kt
@@ -30,6 +30,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import com.programmersbox.kmpuiviews.presentation.components.BackButton
 import com.programmersbox.kmpuiviews.presentation.components.OtakuHazeScaffold
+import com.programmersbox.kmpuiviews.utils.DownloadAndInstallStatus
 import com.programmersbox.uiviews.checkers.DownloadAndInstallState
 import org.koin.androidx.compose.koinViewModel
