+++ UIViews/src/main/java/com/programmersbox/uiviews/presentation/settings/updateprerelease/PrereleaseViewModel.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.programmersbox.uiviews.presentation.settings.downloadstate.DownloadAndInstallStatus
+import com.programmersbox.kmpuiviews.utils.DownloadAndInstallStatus
 import com.programmersbox.uiviews.presentation.settings.downloadstate.DownloadStateRepository
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
